♻️ [Refactoring]: #295 config.rs（約2000行）をドメイン親ファイル規約に従い分割
子フォルダを持つドメイン親ファイルは「pub mod 列挙 + pub use のみ」とする
モジュール記法に config.rs を準拠させる。約2000行同居していたコアスキーマ型・
I/O インフラ・2つのレジストリ集約を以下の子モジュールへ move する（挙動・公開
API・serde 表現は不変、純粋な内部再配置）。

- config/core.rs           — Config / Column / ColumnColor / CardOrder などコア型 +
                              GUIDE.md 生成 / update_columns 純粋計算 /
                              plan_update_card_order / UpdateCardOrderPlanError /
                              build_config_from_statuses / clean_card_order /
                              validate_unique_column_names
- config/migration.rs      — MigrationError / migrate_config
- config/load.rs           — load_or_default / LoadConfigError + atomic write
                              インフラ（ConfigWriter / FsConfigWriter /
                              write_atomic_to_path / backup / cleanup）
- config/label_registry.rs — LabelRegistry / LabelDefinition / LabelColor /
                              LabelGroup / plan_* / store
- config/milestone_registry.rs — MilestoneRegistry / MilestoneDefinition /
                              MilestoneState / plan_* / store

config.rs は pub mod 列挙 + pub use 再エクスポートのみに縮小し、crate::config::X の
公開 API パスを完全に維持する。既存テスト（config_tests / label_registry_tests /
milestone_registry_tests）は親モジュールの子のまま据え置き、super::* 解決のため
コア型・crate 内部 helper・標準 import を親へ取り込む。CLAUDE.md のディレクトリ
構成も新レイアウトへ追従。

main 側の #296（コメント整理）/ #292（update_card_order の plan + effect 2 段化、
plan_update_card_order aggregate メソッド追加）を取り込むため origin/main を起点に
分割を再適用。cargo fmt / clippy --workspace --all-targets（警告ゼロ）/
test --workspace（1010 + 135 件、分割前と同数）すべて green。
